web/xexe/challenge/app/requirements.txt
flask
python-dotenv
uwsgi
lxml
